Draft: split wire related tools from Draft.py
Line Polyline BezCurve BSpline
.


.


.
